Refactor capturing to grab default font size
